requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
-e git+https://bitbucket.org/maykinmedia/django-digid-eherkenning.git@a43e36136fe4a80217508343d60a1e0cc0e6522a#egg=digid_eherkenning
    # via -r requirements/base.in
-e git+https://github.com/maykinmedia/python3-saml.git@984fa09a7861f3b9eaf31b27a7f1eafc14dd9031#egg=python3_saml
    # via -r requirements/base.in
attrs==20.3.0
    # via jsonschema
certifi==2020.6.20
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.14.3
    # via cryptography
chardet==3.0.4
    # via requests
connect-sdk-python3==3.20.0
    # via -r requirements/base.in
cryptography==3.1.1
    # via pyopenssl
defusedxml==0.6.0
    # via
    #   digid-eherkenning
    #   python3-saml
django-admin-index==1.4.0
    # via -r requirements/base.in
django-appconf==1.0.4
    # via django-axes
django-axes==5.5.0
    # via -r requirements/base.in
django-choices==1.7.1
    # via
    #   -r requirements/base.in
    #   digid-eherkenning
    #   zgw-consumers
django-compat==1.0.15
    # via
    #   django-hijack
    #   django-hijack-admin
django-hijack-admin==2.1.10
    # via -r requirements/base.in
django-hijack==2.1.10
    # via
    #   -r requirements/base.in
    #   django-hijack-admin
django-ipware==3.0.1
    # via django-axes
django-ordered-model==3.4.1
    # via django-admin-index
django-redis==4.12.1
    # via -r requirements/base.in
django-relativedelta==1.1.1
    # via zgw-consumers
django-rosetta==0.9.4
    # via -r requirements/base.in
django-solo==1.1.3
    # via
    #   -r requirements/base.in
    #   zgw-consumers
django==2.2.19
    # via
    #   -r requirements/base.in
    #   digid-eherkenning
    #   django-appconf
    #   django-axes
    #   django-choices
    #   django-redis
    #   django-relativedelta
    #   django-rosetta
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   zgw-consumers
djangorestframework-camel-case==1.2.0
    # via -r requirements/base.in
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   drf-nested-routers
    #   drf-spectacular
drf-nested-routers==0.93.3
    # via -r requirements/base.in
drf-spectacular==0.14.0
    # via -r requirements/base.in
elastic-apm==5.8.1
    # via -r requirements/base.in
faker==7.0.1
    # via zgw-consumers
gemma-zds-client==1.0.0
    # via zgw-consumers
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
isodate==0.6.0
    # via python3-saml
jsonschema==3.2.0
    # via drf-spectacular
lxml==4.5.2
    # via xmlsec
pillow==8.2.0
    # via -r requirements/base.in
polib==1.1.0
    # via django-rosetta
psycopg2==2.8.6
    # via -r requirements/base.in
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via gemma-zds-client
pyopenssl==19.1.0
    # via -r requirements/base.in
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   django-relativedelta
    #   faker
python-decouple==3.3
    # via -r requirements/base.in
python-dotenv==0.14.0
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   django
pyyaml==5.3.1
    # via
    #   drf-spectacular
    #   gemma-zds-client
redis==3.5.3
    # via django-redis
requests-mock==1.8.0
    # via zgw-consumers
requests-toolbelt==0.9.1
    # via connect-sdk-python3
requests==2.24.0
    # via
    #   connect-sdk-python3
    #   django-rosetta
    #   gemma-zds-client
    #   requests-mock
    #   requests-toolbelt
    #   zgw-consumers
sentry-sdk==0.17.3
    # via -r requirements/base.in
six==1.15.0
    # via
    #   cryptography
    #   django-choices
    #   django-compat
    #   django-rosetta
    #   isodate
    #   jsonschema
    #   pyopenssl
    #   python-dateutil
    #   requests-mock
sqlparse==0.3.1
    # via django
text-unidecode==1.3
    # via faker
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.25.10
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements/base.in
xmlsec==1.3.8
    # via python3-saml
zgw-consumers==0.15.2
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
